Phillipe's file uploads: Part 5A_CPC
This file takes reads a file named input.csv. Each class will be processed and printed separately on a graph. This is only the first part of task 5.A. Ideally the user would have buttons to choose which graphs to view. When the button is clicked, the main screen would display the figure of graphs and also print the CPC values on the side of the screen